requirements.txt
Django
djangorestframework
drf-spectacular
python-dateutil
Pillow
reportlab
djangorestframework-simplejwt
pyotp
qrcode
django-modeltranslation
